siteMultiplicity can return integer.

+++ src/diffpy/srreal/ObjCrystStructureAdapter.hpp
@@ -63,7 +63,7 @@ class ObjCrystStructureAdapter : public StructureAdapter
         virtual const R3::Vector& siteCartesianPosition(int idx) const;
         virtual double siteOccupancy(int idx) const;
         virtual bool siteAnisotropy(int idx) const;
-        virtual double siteMultiplicity(int idx) const;
+        virtual int siteMultiplicity(int idx) const;
         virtual const R3::Matrix& siteCartesianUij(int idx) const;
         virtual const std::string& siteAtomType(int idx) const;
         const Lattice& getLattice() const;
